chore(onboarding): moves context hook to hooks dir
